Create HasPairWithSum.py
Returns true if there's a pair in an input list that has a sum equal to the input sum. Returns false for empty lists, lists with one element, or lists with no pairs that add up to the sum. I use set instead of list for checking for compliment numbers since its search time is constant compared to list's linear search time.